crawlers/telegram/requirements.txt
altgraph==0.17.2
async-lru==1.0.2
autopep8==1.6.0
et-xmlfile==1.1.0
flake8==4.0.1
macholib==1.15.2
mccabe==0.6.1
modulegraph==0.19.2
numpy==1.22.0
openpyxl==3.0.9
pandas==1.3.5
py2app==0.26.1
pyaes==1.6.1
pycodestyle==2.8.0
pyflakes==2.4.0
Pyrogram==1.2.20
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
six==1.16.0
TgCrypto==1.2.2
toml==0.10.2
